Extract MAPEstimator from MCMC.
